[Project P7] Refactor: extract terrain, integration hook, JSDoc, dead code cleanup
- Extract terrain components into GalaxyTerrain.tsx (391 lines)
- Extract useGalaxyTDIntegration hook from tetris.tsx
- Add JSDoc to all public hooks and context providers
- Remove dead code (unused GalaxyTDPlayerConfig, stale terrainPhase guards)
- GalaxyScene.tsx reduced from 1244 to 870 lines
