+++ src/components/rhythmia/tetris/galaxy-types.ts
@@ -69,14 +69,6 @@ export interface GalaxyGate {
 export type RingTDPhase = 'build' | 'wave' | 'won' | 'lost';
 
 // Full galaxy TD state snapshot
-// Player configuration for multiplayer mode
-export interface GalaxyTDPlayerConfig {
-    playerId: string;
-    controlledSides: GalaxyRingSide[];
-    color: string;
-}
-
-// Full galaxy TD state snapshot
 export interface GalaxyTDState {
     phase: RingTDPhase;
     gold: number;
